Update Babble version and add configurable slow-heartbeat
Using the latest develop version of Babble with a cleaner way
of disabling the API service. Added a slowHeartbeat property
to BabbleConfig to control the frequency of gossip when there
is nothing to gossip about. It used to be hardcoded to 1s,
which made apps seem unresponsive. The default value is 10ms,
like the default heartbeat.