Process InternalTransactions and return block in commitHandler.
The babble onCommit used to return only the stateHash resulting from
applying the block's transactions. The InternalTransactions where
ignored. Now InternalTransactions are decoded, and processed to
produce InternalTransactionReceipts. The stateHash and receipts are
populated directly in the block which is then returned by the
onCommit handler.

WIP:
* InternalTransactions are systematically accepted. We need a
hook to the AppState, similar to applyTransactions, to process them
programmatically.

* Joining/leaving doesn't always work in the sample app.