Add StateChangedHandler to Babble node constructor
The constructor of the Babble-Core node takes an additional parameter
for the StateChangedHandler. This handler is called by Babble-Core
when the Babble node transitions from one state to another ( ex:
Babbling, Joining, Shutdown, Suspended).

TODO: Implement proper hooks with the application to enable developers
to implement a StateChangedHandler, which is currently hard-coded to
log the new state