Stop service when back button is pressd in BaseConfigActivity
If an archive is loaded and the user presses the back button, the
BaseConfigActivity will be destroyed. If in the meantime the service transitions
into the ARCHIVE state, when the app is re-opened the service won't be in the
STOPPED or ARCHIVE-INIT state which the BaseConfigActivity expects. To fix this
the BabbleService now has an additional "stop" method which forces the node to
shutdown and restarts the state. This method is called when the back button is
pressed from the BaseConfigActivity