Ensure TabFragment is shown after returning from Group
Previously when back was pressed from the Sample app's ChatActivity the
BaseConfigActivity would pick up where it left off - either displaying the
JoinGroupFragment or the NewGroupFragment. This change means it will now
display the TabFragment. This is handled in the onStart method, which means
care needs to be taken to ensure this Fragment replacement only occurs when
starting after onJoin or onNew and not after a config change.