Prevent app from crashing if WiFi is turned off
When starting a new group the app would crash if any exceptions were thrown when
starting the Babble service. This change catches IllegalArgumentExceptions which
can be caused by the port being in use or if the WiFi is turned off. The port
may be in use because we are waiting fo a leave to complete. Inside the catch a
dialog is displayed suggesting to the user that they wait a few seconds and
check their WiFi connection.