Improvements and fixes to archive display
Improve and fix archive display:

 - When long clicking an archived group, every n'th group would be
   selected. Change to only select the clicked group.
 - Maintain group selections through configuration changes.
 - Remove Contextual Action Bar when changing to another tab

A side effect of these changes has been to simplify the
ArchivedGroupsFragment. The archive fragment does not wait for the
archive to load anymore. This has simplified the ArchivedGroupsFragment,
the BaseConfigActivity and the BabbleService but requires some
additional checking in the ChatActity and when returning to the
BaseConfigActivity from a chat.