Create Gradle plugin for fetching Babble Core
The code to fetch Babble Core was written directly in the babble build
script. This has been moved into a Gradle plugin. A small modification
has been made to ensure that a failure to fetch the artifacts on a
previous run is recognised on the current run so that another attempt at
fetching is made.