Add BabbleService2 Android Service
Add an Android Service, BabbleService2, which replicates much of the
functionality of the BabbleService singleton class. For creating a new
group and opening an archive, the new BabbleService2 is used. For
joining a group the original BabbleService is used. Within the sample
app creating a new group or opening an archive uses a new ChatActivity2.
The intention is to remove BabbleService and ChatActivity once the
discovery process has been refactored.