Fix bug when reloading tabs
When back is pressed when viewing an archive the TabsFragment is re
created using the default values for the flags which control which tabs
to show. The underlying issue is that the TabsFragment should not be
recreated when back is pressed. This change is a band aid to mask this
problem.