Stop displaying incorrect date on notifications
Notification messages are created and timestamped at the point a client
processes a join request. This differs to regular messages which are
stored in the db. This causes jumps in time when displaying archives.
This commit fixes this by ignoring notifications when deciding whether
to display the date in the chat history.