Removed calls to BabbleConstants in ConfigManager
Hard coded various parameters that were pulled from the BabbleConstants
class. Since the ConfigManager can be used in isolation from the UI
components, it should be configured using method calls. File names
have not been made configurable since they cannot be changed without
changing the go code. The babble directory name is not configurable,
this can be changed if necessary.

The BabbleConstants class is no longer used however it remains in the
code base as it will be used in future for configuring UI components.